# api_gateway/docs/swagger.yaml
definitions:
  conversationhandler.createConversationRequest:
    properties:
      second_user_id:
        type: string
    type: object
  dto.ConversationMessageDTO:
    properties:
      created_at:
        type: string
      message_body:
        type: string
      sender_id:
        type: string
    type: object
  dto.ConversationsDTO:
    properties:
      id:
        type: integer
      last_message:
        $ref: '#/definitions/dto.ConversationMessageDTO'
      user:
        $ref: '#/definitions/userEntity.PublicUser'
    type: object
  dto.MessageDTO:
    properties:
      created_at:
        type: string
      message_body:
        type: string
      sender_id:
        type: string
      user:
        $ref: '#/definitions/dto.MessageUserDTO'
    type: object
  dto.MessageUserDTO:
    properties:
      id:
        type: string
      username:
        type: string
    type: object
  jwtmanager.ReturnTokenType:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  messagehandler.createMessageRequestBody:
    properties:
      conversation_id:
        type: integer
      text:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      status_code:
        type: integer
    type: object
  userEntity.PublicUser:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  userEntity.Role:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - Admin
    - Regular
  userEntity.UpdateUserRequest:
    properties:
      code:
        type: string
      email:
        type: string
      enabled:
        type: boolean
      role:
        $ref: '#/definitions/userEntity.Role'
      username:
        type: string
    type: object
  userhandler.checkCodeReq:
    properties:
      code:
        type: string
      id:
        type: string
    type: object
  userhandler.refreshRequest:
    properties:
      refreshToken:
        example: your_refresh_token_here
        type: string
    type: object
  userhandler.requestBody:
    properties:
      email:
        type: string
    type: object
  userhandler.resp:
    properties:
      id:
        type: string
    type: object
info:
  contact: {}
  title: API Gateway
  version: "1.0"
paths:
  /api/v1/conversations:
    get:
      consumes:
      - application/json
      description: Get conversations that belongs to user
      parameters:
      - description: Skip items count
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ConversationsDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get conversations
      tags:
      - conversations
    post:
      consumes:
      - application/json
      description: Create conversation
      parameters:
      - description: Request body containing user_id and second_user_id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/conversationhandler.createConversationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create conversation
      tags:
      - conversations
  /api/v1/messages:
    post:
      consumes:
      - application/json
      description: Create message
      parameters:
      - description: Request body containing conversation_id and text of the message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/messagehandler.createMessageRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create message
      tags:
      - messages
  /api/v1/messages/{conversation_id}:
    get:
      consumes:
      - application/json
      description: Get messages by conversation id
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MessageDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get messages
      tags:
      - messages
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/userEntity.PublicUser'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: User with optional fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userEntity.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/userEntity.PublicUser'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /api/v1/users/{id}:
    get:
      consumes:
      - application/json
      description: Get user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/userEntity.PublicUser'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Get user
      tags:
      - users
  /api/v1/users/auth:
    post:
      consumes:
      - application/json
      description: Authentication by username and email
      parameters:
      - description: Request body containing username and email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userhandler.requestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/userhandler.resp'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Authentication
      tags:
      - users
  /api/v1/users/check-code:
    post:
      consumes:
      - application/json
      description: Check code with id and code
      parameters:
      - description: Request body containing ID and Code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userhandler.checkCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jwtmanager.ReturnTokenType'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Check code
      tags:
      - users
  /api/v1/users/refresh:
    post:
      consumes:
      - application/json
      description: Refresh token
      parameters:
      - description: Request body containing refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userhandler.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jwtmanager.ReturnTokenType'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Refresh
      tags:
      - users
  /api/v1/users/search:
    get:
      consumes:
      - application/json
      description: Search users
      parameters:
      - description: Search query
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/userEntity.PublicUser'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Search users
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"